// src/app/utils/models/form.ts
export class Form {
  id: number;
  beneficiary_id: number;
  form_type: number;
  date_from: string;
  date_to: string;
  property_name: string;
  client_id: number;
  client_type: number;
  client_checking_account: string;
  region_id: number;
  quantity: number;
  insurance_cost: number;
  insurance_sum: number;
  anti_fire_stuff: number;
  security_stuff: number;
  payment_type: number;
  payment_currency: number;
  insurer_id: number;
}